Distribute the tests to the appropriate test classes
These are not tests of the module itself, but simply that other components behave correctly in modules
