+++ src/test/java/spoon/test/compilationunit/TestCompilationUnit.java
@@ -40,10 +40,13 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
@@ -131,6 +134,22 @@ public class TestCompilationUnit {
 			// do nothing
 		}
 	}
+
+	@Test
+	public void testCompilationUnitInNamedModuleHasDeclaredTypes() {
+		// contract: Compilation units in named modules should have the expected declared types
+
+		final Launcher launcher = new Launcher();
+		launcher.getEnvironment().setComplianceLevel(9);
+		launcher.addInputResource("./src/test/resources/spoon/test/module/simple_module_with_code");
+		launcher.buildModel();
+
+		CtType<?> expectedType = launcher.getFactory().Type().get("fr.simplemodule.pack.SimpleClass");
+		CtCompilationUnit cu = launcher.getFactory().CompilationUnit().getOrCreate(expectedType);
+
+		assertThat(cu.getDeclaredTypes(), equalTo(Collections.singletonList(expectedType)));
+	}
+
 
 	@Test
 	public void testCompilationUnitSourcePosition() throws IOException {
